third_party/blink/web_tests/platform/win10/external/wpt/css/css-fonts/cjk-kerning-expected.txt

This is a testharness.js-based test.
[FAIL] expected match: .kernON .cjk vs .paltONkernON .cjk
  assert_equals: expected 316 but got 384
[FAIL] expected mismatch: .kernOFF .cjk vs .kernON .cjk
  assert_not_equals: got disallowed value 384
Harness: the test ran to completion.
